Correccion de nombres

+++ entrada.cc
@@ -29,7 +29,7 @@ int Factorial (int number) {
 
 int main () {
   std :: cout << " Introduzca el numero de factoriales a calcular : ";
-  int limit ;
+  double limit ;
   std :: cin >> limit ;
   for (int i = 1; i <= limit ; ++ i ) {
     std :: cout << i << "! = " << ( double ) Factorial ( i ) << std :: endl ;
